+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Update table component documentation
 * Fix some translations
 * Update st-page-title doc
+* Update radio menu component documentation
+
 
 
 ## 1.3.0 (Abril 4, 2017)